:art: Move GameFile into new FileSystem namespace This namespace will contain everything about the virtual FileSystems. Issue #5
