+++ libgame/FileSystem/GameFile.cs
@@ -18,7 +18,7 @@
 //
 //  You should have received a copy of the GNU General Public License
 //  along with this program.  If not, see <http://www.gnu.org/licenses/>.
-namespace Libgame
+namespace Libgame.FileSystem
 {
     using System;
     using System.Collections.Generic;
