+++ libgame/libgame.csproj
@@ -8,6 +8,9 @@
     <RootNamespace>Libgame</RootNamespace>
     <AssemblyName>libgame</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
+    <ProductVersion>8.0.30703</ProductVersion>
+    <SchemaVersion>2.0</SchemaVersion>
+    <ReleaseVersion>1.0</ReleaseVersion>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>true</DebugSymbols>
@@ -40,7 +43,6 @@
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="FileContainer.cs" />
-    <Compile Include="GameFile.cs" />
     <Compile Include="GameFolder.cs" />
     <Compile Include="Extensions\NumericExtensions.cs" />
     <Compile Include="Extensions\StringExtensions.cs" />
@@ -61,6 +63,7 @@
     <Compile Include="FileFormat\BinaryFormat.cs" />
     <Compile Include="PluginManager.cs" />
     <Compile Include="FileFormat\IPartialConverter.cs" />
+    <Compile Include="FileSystem\GameFile.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <ItemGroup>
@@ -68,5 +71,6 @@
     <Folder Include="IO\" />
     <Folder Include="Utils\" />
     <Folder Include="FileFormat\" />
+    <Folder Include="FileSystem\" />
   </ItemGroup>
 </Project>
